% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhoplus_PP.R
\name{rhominus_PP}
\alias{rhominus_PP}
\title{Computes the non-parametric estimator of the multivariate Spearman's rho measure of average lower orthant dependence proposed by Pérez and Prieto-Alaiz (2016).}
\usage{
rhominus_PP(X)
}
\arguments{
\item{X}{nxd matrix of the data}
}
\value{
value of the coefficient
}
\description{
Computes the non-parametric estimator of the multivariate Spearman's rho measure of average lower orthant dependence proposed by Pérez and Prieto-Alaiz (2016).
}
\examples{
library(copula)
set.seed(111)
mv.Cl.NN <- mvdc(claytonCopula(dim = 3, iRho(claytonCopula(),0.7)), c("norm", "norm", "norm"),
list(list(mean = 0, sd =2), list(mean = 0, sd =2), list(mean = 0, sd =2)))
x.sample <- rMvdc(10000, mv.Cl.NN)
rhominus_PP(x.sample)
}
