% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhoplus_PP.R
\name{rhoplus_GNR}
\alias{rhoplus_GNR}
\title{Computes the tie-corrected non-parametric estimator of the multivariate Spearman's rho measure of average upper orthant dependence proposed by Genest et al. (2013).}
\usage{
rhoplus_GNR(data)
}
\arguments{
\item{data}{nxd matrix of the data}
}
\value{
value of the coefficient
}
\description{
Computes the tie-corrected non-parametric estimator of the multivariate Spearman's rho measure of average upper orthant dependence proposed by Genest et al. (2013).
}
\examples{
library(copula)
set.seed(111)
mv.Cl.NN <- mvdc(claytonCopula(dim = 3, iRho(claytonCopula(),0.7)), c("norm", "norm", "norm"),
list(list(mean = 0, sd =2), list(mean = 0, sd =2), list(mean = 0, sd =2)))
x.sample <- rMvdc(10000, mv.Cl.NN)
rhoplus_GNR(x.sample)
}
